Move some functions to helper module
